backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
pydantic==2.9.1
sqlalchemy==2.0.32
alembic==1.13.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
shap==0.46.0
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
